pyproject.toml
[project]

name = "refractiveindex2"
version = "v0.0.4"
description = "A python interface to the refractiveindex.info database."
keywords = []
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}

authors = [
    {name = "Martin Schubert", email = "mfschubert@gmail.com"},
]
maintainers = [
    {name = "Martin Schubert", email = "mfschubert@gmail.com"},
]

dependencies = [
    "numpy",
    "pyyaml",
    "requests",
    "types-PyYAML",
    "types-requests",
]

[project.optional-dependencies]
tests = [
    "parameterized",
    "pytest",
    "pytest-cov",
]
dev = [
    "bump-my-version",
    "darglint",
    "isort",
    "mypy",
    "pre-commit",
]

[tool.setuptools.package-data]
"refractiveindex2" = ["py.typed"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.pydocstyle]
inherit = false
convention = "google"
